requirements.txt
Django==3.1.3
djangorestframework==3.12.2
docutils==0.16
nltk==3.5
nose==1.3.7
numpy==1.18.5
pandas==1.1.3
Pillow==8.0.1
pycodestyle==2.6.0
scikit-learn==0.23.2
tensorflow==2.3.1
tflearn==0.5.0
# pytorch==1.7.0
# torchvision==0.8.1